+++ examples/resources/morpheus_password_option_type/resource.tf
@@ -1,14 +1,19 @@
 resource "morpheus_password_option_type" "tf_example_password_option_type" {
   name                     = "tf_example_password_option_type"
   description              = "Terraform password option type example"
-  field_name               = "test1"
+  labels                   = ["demo", "terraform"]
+  field_name               = "tfPasswordExample"
   export_meta              = true
-  dependent_field          = "servicemsh"
-  visibility_field         = "demotestin"
+  dependent_field          = "dependent_example"
+  visibility_field         = "visibility_example"
+  require_field            = "require_example"
+  show_on_edit             = true
+  editable                 = true
   display_value_on_details = true
   field_label              = "numbers"
   placeholder              = "fewf"
   default_value            = "testing"
   help_block               = "fiwefw"
   required                 = true
+  verify_pattern           = "a\\D{4}"
 }